Code cleanup on hyperv-attachment + better EPT implementation + deep copy
Code in the hyperv-attachment has been massively cleaned up and given improved const-correctness.

EPT implementation now uses a second, hook EPTP. It works as such:

- target page is set to --x (not readable, not writable, but is executable) with the shadow page frame number in the Hyper-V EPTP
- target page is set to rw- (readable, writable, but not executable) with original page frame number in the hook EPTP.

When a read/write happens under the Hyper-V EPTP of the hooked page, the EPTP is changed to hook EPTP, so the read/write will occur on the original page frame number (this masks the true contents of the hooked page). The opposite happens for when it is executed.

Now, we deep copy the Hyper-V EPTP/NCR3 on both Intel and AMD, so their hook SLAT CR3s are 1:1 with the Hyper-V SLAT CR3.